linkedbench-iot/linkedbench-iot/requirements.txt
# LinkedBench IoT System Requirements

# GPIO and Hardware
RPi.GPIO>=0.7.1
smbus2>=0.4.2

# MQTT Communication
paho-mqtt>=1.6.1

# REST API
Flask>=2.3.0
Flask-CORS>=4.0.0

# HTTP requests (for ThingSpeak, etc.)
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2
